% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.files.r
\name{get.files}
\alias{get.files}
\title{This function returns the paths of all files a given folder}
\usage{
get.files(folder.path, recursive = FALSE)
}
\arguments{
\item{folder.path}{The path of the folder}

\item{recursive}{Search for files in subfolders. Defaults to FALSE}
}
\value{
A list of file paths
}
\description{
This function returns the paths of all files a given folder
}
\examples{
get.files("./path/to/folder")
}

