% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.folder.r
\name{read.files}
\alias{read.files}
\title{This function reads a list of folders using the read function of your choice}
\usage{
read.files(file.list, read.function, ...)
}
\arguments{
\item{file.list}{A list of file paths}

\item{read.function}{A read function. Does not have a default. Possible suggestions are read.csv, read.table, fread, etc}
}
\value{
This function returns the object (data.frame, data.table, etc) of your choice containing data from all files specified
}
\description{
This function reads a list of folders using the read function of your choice
}
\examples{
read.files(c("file1.csv", "file2.csv", read.function=read.table, col.names=c("a","b","c"), colClasses=c("character","character","numeric"))
}

