% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.folder.r
\name{read.folder}
\alias{read.folder}
\title{Reads an entire folder using the read function of your choice}
\usage{
read.folder(folder.path, read.function = read.csv, recursive = FALSE,
  exclude = NULL, ...)
}
\arguments{
\item{folder.path}{The path of the folder}

\item{read.function}{The function used to read the files. Do not surround in quotes. Deafults to read.csv}

\item{recursive}{Whether we should look inside subfolders. Defaults to FALSE}

\item{exclude}{Paths of all files we would like to exclude from reading (for example README files)}
}
\value{
This function returns the object (data.frame, data.table, etc) of your choice containing data from all files in the specified folder
}
\description{
Reads an entire folder using the read function of your choice
}
\examples{
read.folder("./path/to/folder/", read.function=fread)
read.folder("./path/to/folder/", exclude="./path/to/folder/README.txt", recursive=TRUE, header=FALSE)
}

